#20 refactor: move workspace-related code into separate package
