+++ graph_explorer/graph_explorer/templates/base.html
@@ -16,11 +16,6 @@
 
     </div>
     <div class="left">
-        <ul>
-            <li>
-                <a href="{% url 'index' %}">Home</a>
-            </li>
-        </ul>
         <p>Tree layout</p>
         <div class="tree-container">
             {{ tree_view }}
